ESQL: Remove esql-core's sort
This moves the parts of `Sort` that we use out of esql-core and removes
the rest. And the subclass.
